REFACTOR code
