+++ css/style.css
@@ -9,6 +9,10 @@ main {
   /* W3C, IE 10+/ Edge, Firefox 16+, Chrome 26+, Opera 12+, Safari 7+ */
 }
 
+input[type='text'] {
+  margin-top: 15px;
+}
+
 .hide {
   display: none;
 }
@@ -130,6 +134,7 @@ main {
   margin: 0 auto;
   left: 33%;
   width: 33%;
+  border-radius: 12px;
   top: 12vh;
   background-color: #fff;
   padding: 1rem 2rem 2rem;
