+++ src/project.js
@@ -1,0 +1,8 @@
+// creates project object
+
+const newProject = (title) => ({
+  title: title,
+  tasks: []
+})
+
+export default newProject
