+++ src/task.js
@@ -1,0 +1,12 @@
+// creates task object
+
+const newTask = (title, description = '', dueDate = '', priority = '', completed = false, project = 'default') => ({
+  title: title,
+  description: description,
+  dueDate: dueDate,
+  priority: priority,
+  completed: completed,
+  project: 'default'
+})
+
+export default newTask
